Renaming container
